Update Go versions in .travis.yml
Specifying a version as e.g. "1.6.x" will cause the latest minor
revision of that version to be used, which is probably preferable to
avoid running into unexpected compiler bugs.
Since minor revisions are always supposed to only fix critical problems
in in Go releases and should not change the API, this should always be fine.

Additionally, replace 1.7 with a reference to the latest stable release
of Go 1.

Since godo/context still maintains support for pre 1.7, however, Go 1.6
is retained so that tests are still run against it.